Create nix flake and add python script

+++ compile.sh
@@ -1,3 +1,7 @@
+#!/bin/bash
+if [ ! -f src/SimConfig.hpp ] || [ "$1" = "--configure" ]; then
+    python3 configure.py || exit 1
+fi
 rm -rf build
 cmake -S . -B build -DCMAKE_EXPORT_COMPILE_COMMANDS=ON
 cmake --build build
